// src/app/enums/feature.enum.ts
/* eslint-disable @typescript-eslint/naming-convention */
export enum FeatureEnum {
  BANDA_LARGA_LENTA = 1,
  BANDA_LARGA_SEM_CONEXAO = 2,
  BANDA_LARGA_INTERMITENTE = 3,
  FIBRA_DISPOSITIVOS_CONECTADOS = 4,
  BANDA_LARGA_DISPOSITIVOS_CONECTADOS = 5,
  FIBRA_INTENSIDADE_SINAL = 6,
  BANDA_LARGA_TROCA_SENHA = 7,
  BANDA_LARGA_TROCA_NOME = 8,
  FIBRA_CONTROLE_REMOTO = 9,
  FIBRA_RECURSOS_AVANCADOS = 10,
  FIBRA_GUIA_PROGRAMACAO = 11,
  SMARTV_WIFI = 12,
  MELHORAR_WIFI = 13,
  FAZER_LIGACAO = 14,
  FIBRA_LENTA = 15,
  FIBRA_SEM_CONEXAO = 16,
  FIBRA_INTERMITENTE = 17,
  FIBRA_PROBLEMAS_TV = 18,
  FIBRA_LINHA_MUDA = 19,
  FIBRA_NAO_FAZ_CHAMADA = 20,
  FIBRA_NAO_RECEBE_CHAMADA = 21,
  FIBRA_NAO_FAZ_NEM_RECEBE_CAHAMDA = 22,
  FIBRA_TELEFONE_COM_RUIDO = 23,
  FIXO_LINHA_MUDA = 24,
  FIXO_NAO_FAZ_CHAMADA = 25,
  FIXO_NAO_RECEBE_CHAMADA = 26,
  FIXO_NAO_FAZ_NEM_RECEBE_CAHAMDA = 27,
  FIXO_TELEFONE_COM_RUIDO = 28,
  TV_AUSENCIA_SINAL = 29,
  TV_TELA_PRETA = 30,
  TV_CARTAO_INCOMPATIVEL = 31,
  TV_SERVICO_INATIVO = 32,
  FIXO_LINHA_CRUZADA = 33,
  FIBRA_TROCA_SENHA = 34,
  FIBRA_TROCA_NOME = 35,
  BANDA_LARGA_AGENDAMENTO = 36,
  DIAGNOSTICO_HOME = 37,
  FIBRA_COMPATIBILIDADE_VELOCIDADES = 38,
  FIBRA_TV_TELA_PRETA = 39,
  OI_PRODUCTS = 40,
  TV_CONTROLE_REMOTO = 41,
  FIBRA_CONFIG_ANTENA_MODEM = 42,
  FIBRA_WELCOME_KIT = 43,
  RETIRADA_EQUIPAMENTO_FIBRA = 44,
  DEFEITO_NO_CONTROLE = 45,
  FIBRA_TV_IMAGEM_TOMADA= 46,
  FIBRA_TV_IMAGEM_LOGO_OI= 47,
  FIBRA_NETFLIX= 48,
  FIBRA_DIAGNOSTICO_HOME = 49,
}